+++ ohsome_quality_api/geodatabase/select_building_area.sql
@@ -8,4 +8,4 @@ FROM
     eubucco,
     bpoly
 WHERE
-    ST_Intersects (eubucco.geom, bpoly.geom);
+    ST_Intersects (eubucco.centroid, bpoly.geom);