+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 
 - building-comparison: no quality estimation for areas with strong difference to reference data ([#739])
 - test(db): add missing mock for getting coverage from database in tests ([#739])
+- building-comparison: request EUBUCCO database now 
 
 [#739]: https://github.com/GIScience/ohsome-quality-api/pull/739